locale: update locales to include all Sugar langs
update lang script README
