+++ package.json
@@ -61,10 +61,14 @@
     "topojson": "^1.6.26"
   },
   "devDependencies": {
+    "cldr": "^4.5.0",
+    "cldr-data": "^31.0.2",
+    "cldrjs": "^0.5.0",
     "clean-css": "^3.4.8",
     "codecov": "^2.2.0",
     "extend-tape": "^1.2.0",
     "faucet": "0.0.1",
+    "globalize": "^1.3.0",
     "istanbul": "^0.4.5",
     "jsdom": "^11.0.0",
     "jshint": "^2.9.4",
@@ -88,6 +92,7 @@
     "stylelint-config-standard": "^16.0.0",
     "tape": "^4.6.2",
     "tapes": "^4.1.0",
+    "twitter_cldr": "^2.1.1",
     "uglify-js": "2.6.1"
   }
 }
