+++ src/scripts/lang/cldr-locales.json
@@ -1,0 +1,41 @@
+[
+  ["ar_SA", "ar", "SAR", "ar-SA", "Arabic (Saudi Arabia)"],
+  ["bg_BG", "bg", "BGN", "bg", "Bulgarian (Bulgaria)"],
+  ["ca_ES", "ca", "EUR", "ca", "Catalan (Catalonia)"],
+  ["cs_CZ", "cs", "CZK", "cs", "Czeck (Czechia)"],
+  ["da_DK", "da", "DKK", "da", "Danish (Denmark)"],
+  ["de_DE", "de", "EUR", "de", "German (Germany)"],
+  ["el_EL", "el", "EUR", "el", "Greek (Greece)"],
+  ["en_UK", "en-GB", "EUR", "en-GB", "English (Great Britan)"],
+  ["en_US", "en-US", "USD", "en", "Engligh (US)"],
+  ["es_ES", "es-ES", "EUR", "es", "Spanish (Spain)"],
+  ["es_LA", "es-419", "MXN", "es-419", "Spanish (Latin America)"],
+  ["et_EE", "et", "EUR", "et", "Estonian (Estonia)"],
+  ["fi_FI", "fi", "EUR", "fi", "Finish (Finland)"],
+  ["fr_FR", "fr", "EUR", "fr", "French (France)"],
+  ["he_IL", "he", "ILS", "he", "Hebrew (Israel)"],
+  ["hr_HR", "hr", "HRK", "hr", "Croatian (Croatia)"],
+  ["hu_HU", "hu", "HUF", "hu", "Hungarian (Hungaria)"],
+  ["it_it", "it", "EUR", "it", "Italian (Italy)"],
+  ["ja_JP", "ja", "JPY", "ja", "Japanese (Japan)"],
+  ["ko_KR", "ko", "KRW", "ko", "Korean (Republic of Korea)"],
+  ["lt_LT", "lt", "EUR", "lt", "Lithuanian (Lithuania)"],
+  ["lv_LV", "lv", "EUR", "lv", "Latvian (Latvia)"],
+  ["nb_NO", "nb", "NOK", "nb", "Norwegian Bokmål (Norway)"],
+  ["nl_NL", "nl", "EUR", "nl", "Dutch (Netherlands)"],
+  ["pl_PL", "pl", "PLN", "pl", "Polish (Poland)"],
+  ["pt_BR", "pt-BR", "BRL", "pt", "Portuguese (Brazil)"],
+  ["pt_PT", "pt-PT", "EUR", "pt-PT", "Portuguese (Portugal)"],
+  ["ro_RO", "ro", "RON", "ro", "Romanian (Romania)"],
+  ["ru_RU", "ru", "RUB", "ru", "Russian (Russian Federation)"],
+  ["sk_SK", "sk", "EUR", "sk", "Slovak (Slovakia)"],
+  ["sq_AL", "sq", "ALL", "sq", "Albanian (Albania)"],
+  ["sr_RS", "sr", "RSD", "sr", "Serbian (Serbia)"],
+  ["sv_SE", "sv", "SEK", "sv", "Swedish (Sweden)"],
+  ["th_TH", "th", "THB", "th", "Thai (Thailand)"],
+  ["tr_TR", "tr", "TRY", "tr", "Turkish (Turkey)"],
+  ["uk_UA", "uk", "UAH", "uk", "Ukrainian (Ukrainia)"],
+  ["zh_CN", "zh", "CNY", "zh", "Traditional Chinese (China)"],
+  ["zh_TW", "zh-Hant", "TWD", "zh-Hant", "Chinese (Taiwan)"],
+  ["zu_ZA", "zu", "ZAR", "zu", "Zulu (South Africa)]"]
+]
