+++ src/scripts/lang/locales.json
